fix(map): Improve modifier handling for IsLetter keys
Rather than having to fix up a whole laundry list of broken modifier
keys in the 'qtcode' field of mappings, I'm just going to hardcode
the behaviour I want this to have, which is that the current modifiers
are applied to IsLetter keys.

This is needed as otherwise you need to specify the correct qtcode (plus
modifiers!) for each IsLetter mapping, and you need one mapping _for all
modifier combinations_, so it gets exhausting and boring very very quickly.

As a part of this change, we remove most IsLetter mappings, since they
are now redundant. There's a number of IsLetter that are still remaining,
with special behaviors attached, but I'm choosing to leave those in
place, because fixing them all and verifying that they have the correct
behaviour is overly time consuming and risky to do at this point.