fix(map): Fix handling of modifiers for ZX keys
Layouts that have customised handling for ZX keys (e.g. Norwegian, which
has a <> on them) hadn't handled the customisation properly. By
specifying the customised key for e.g. Ctrl+Z returning <, this broke
shortcut handling.

This fixes the mappings for those cases so that the shortcuts work as expected.