fix(handler): Widen modifiers to 16 bits, and add ModMeta
This isn't sent - yet - the mappings will be adjusted in a followup.

But the intention is that the HW Opt key will send Qt::MetaModifier, so
that it works the same way as Qt on Mac. Then, we just need to handle
the key being remapped, and we're golden.

I didn't bother mapping a left and right meta, even though for some
keyboards that might be needed, as _we_ don't make use of that, so...
yeah.

Task: APPS-14126